feat: add more checked-unchecked interfaces and borrowing for uints
