+++ manta-crypto/src/constraint.rs
@@ -29,8 +29,7 @@ pub use crate::eclair::{
         mode::{self, Derived, Public, Secret},
         Allocate, Allocator, Const, Constant, Var, Variable,
     },
-    assert::{Assert, AssertEq},
-    bool::{Bool, ConditionalSelect, ConditionalSwap},
+    bool::{Assert, AssertEq, Bool, ConditionalSelect, ConditionalSwap},
     cmp::{Eq, PartialEq},
     ops::{Add, Not, Sub},
     Has, Native,
